Add basic behave test for accessing participant journals
