+++ features/environment.py
@@ -1,13 +1,10 @@
 from splinter import Browser
 from behave_django import environment
-
+from django.conf import settings
 from worth2.main.auth import generate_password
 from worth2.main.tests.factories import (
-    ParticipantFactory, WorthModuleFactory
+    ParticipantFactory, UserFactory, WorthModuleFactory
 )
-
-
-BEHAVE_DEBUG_ON_ERROR = False
 
 
 def before_all(context):
@@ -28,6 +25,9 @@ def before_scenario(context, scenario):
     participant.user.set_password(password)
     participant.user.save()
 
+    # Make a facilitator
+    UserFactory()
+
 
 def after_scenario(context, scenario):
     environment.after_scenario(context, scenario)
@@ -39,6 +39,6 @@ def after_all(context):
 
 
 def after_step(context, step):
-    if BEHAVE_DEBUG_ON_ERROR and step.status == "failed":
+    if settings.BEHAVE_DEBUG_ON_ERROR and step.status == "failed":
         import ipdb
         ipdb.post_mortem(step.exc_traceback)
