+++ features/index.feature
@@ -1,7 +1,7 @@
 Feature: Index Page
   Scenario: Index Page Load
     When I access the url "/"
-    Then I get a 302
+    Then I get a 200
 
   Scenario: Facilitator Sign In
     When I sign in as a facilitator
