+++ features/steps/auth.py
@@ -1,16 +1,18 @@
-from behave import given
+import urlparse
+from behave import given, when
 from django.conf import settings
 from django.contrib.auth import BACKEND_SESSION_KEY, SESSION_KEY
 from django.contrib.auth.models import User
 from django.utils.module_loading import import_module
 from splinter.request_handler.status_code import HttpResponseError
-
-from worth2.main.models import Participant
+from worth2.main.auth import generate_password
+from worth2.main.models import Location, Participant
+from worth2.main.tests.factories import UserFactory
 
 
 def create_pre_authenticated_session(user_type):
     if user_type == 'participant':
-        # TODO: why this work on participants?
+        # TODO: why doesn't this work on participants?
         participant = Participant.objects.first()
         user = participant.user
     else:
@@ -42,3 +44,49 @@ def i_am_signed_in_as_a(context, user_type):
     user, s = create_pre_authenticated_session(user_type)
     b.cookies.add({'name': settings.SESSION_COOKIE_NAME, 'value': s})
     context.user = user
+
+
+@when(u'I sign in as a facilitator')
+def i_sign_in_as_a_facilitator(context):
+    facilitator = UserFactory()
+    facilitator.set_password('test_pass')
+    facilitator.save()
+
+    b = context.browser
+    b.visit(urlparse.urljoin(context.base_url, '/accounts/login/'))
+    b.fill('username', facilitator.username)
+    b.fill('password', 'test_pass')
+    b.find_by_css('.form-signin button[type="submit"]').first.click()
+
+
+@when(u'I sign in as a participant')
+def i_sign_in_as_a_participant(context):
+    i_sign_in_as_a_facilitator(context)
+
+    participant = Participant.objects.first()
+    password = generate_password(participant.user.username)
+    participant.user.set_password(password)
+    participant.user.save()
+
+    location = Location.objects.first()
+
+    b = context.browser
+    b.visit(urlparse.urljoin(context.base_url, '/sign-in-participant/'))
+    b.select('participant_id', participant.pk)
+    b.select('participant_location', location.pk)
+    b.choose('participant_destination', '1')
+    b.choose('session_type', 'regular')
+    b.find_by_css(
+        '.worth-facilitator-sign-in-participant button[type="submit"]'
+    ).first.click()
+
+
+@when(u'I sign in as the facilitator from a session')
+def i_sign_in_as_the_facilitator_from_a_session(context):
+    b = context.browser
+    b.find_by_text('Facilitator').first.click()
+    facilitator = User.objects.filter(is_active=True).first()
+    b.fill('facilitator_username', facilitator.username)
+    b.fill('facilitator_password', 'test')
+    b.find_by_css(
+        '.worth-sign-out-participant button[type="submit"]').first.click()
