+++ features/steps/common.py
@@ -1,45 +1,6 @@
 import urlparse
 from behave import when, then
 
-from worth2.main.auth import generate_password
-from worth2.main.models import Location
-from worth2.main.tests.factories import UserFactory, ParticipantFactory
-
-
-@when(u'I sign in as a facilitator')
-def i_sign_in_as_a_facilitator(context):
-    facilitator = UserFactory()
-    facilitator.set_password('test_pass')
-    facilitator.save()
-
-    b = context.browser
-    b.visit(urlparse.urljoin(context.base_url, '/accounts/login/'))
-    b.fill('username', facilitator.username)
-    b.fill('password', 'test_pass')
-    b.find_by_css('.form-signin button[type="submit"]').first.click()
-
-
-@when(u'I sign in as a participant')
-def i_sign_in_as_a_participant(context):
-    i_sign_in_as_a_facilitator(context)
-
-    participant = ParticipantFactory()
-    password = generate_password(participant.user.username)
-    participant.user.set_password(password)
-    participant.user.save()
-
-    location = Location.objects.first()
-
-    b = context.browser
-    b.visit(urlparse.urljoin(context.base_url, '/sign-in-participant/'))
-    b.select('participant_id', participant.pk)
-    b.select('participant_location', location.pk)
-    b.choose('participant_destination', '1')
-    b.choose('session_type', 'regular')
-    b.find_by_css(
-        '.worth-facilitator-sign-in-participant button[type="submit"]'
-    ).first.click()
-
 
 @when(u'I access the url "{url}"')
 def i_access_the_url(context, url):
@@ -55,6 +16,23 @@ def i_click_the_next_button(context):
 def i_click_the_submit_button(context):
     context.browser.find_by_css(
         '.pagetree-form-submit-area input[type="submit"]').first.click()
+
+
+@when(u'I click the first participant journal button')
+def i_click_the_first_participant_journal_button(context):
+    context.browser.find_by_css(
+        'button[data-target^="#view-participant-journal-modal-"]'
+    ).first.click()
+
+
+@when(u'I click the link "{text}"')
+def i_click_the_link(context, text):
+    context.browser.find_link_by_partial_text(text).first.click()
+
+
+@when(u'I click the button "{text}"')
+def i_click_the_button(context, text):
+    context.browser.find_by_text(text).first.click()
 
 
 @then(u'I see the text "{text}"')
