+++ requirements.txt
@@ -76,4 +76,4 @@ parse-type==0.3.4
 parse==1.6.6
 behave==1.2.5
 behave-django==0.1.4
-splinter==0.7.2
+splinter==0.7.3rc1
